Restructured dashboard/controller/classroom
